requirements.txt
flask==3.1.0
telethon==1.37.0
cryptg==0.4.0
aiofiles==24.1.0
python-dotenv==1.0.1
werkzeug==3.1.0
